movthings into their own direcories becuse I was having an error
