--- backend/internal/server/apis/schedule/analysis_test.go
@@ -1,1 +1,0 @@
-package schedule
