+++ backend/internal/server/apis/schedule/ref/ref.go
@@ -1,12 +1,18 @@
-package schedule
+package ref
 
 import (
 	"fmt"
 	"github.com/gofiber/fiber/v2"
+	"github.com/jak103/powerplay/internal/server/apis"
 	"github.com/jak103/powerplay/internal/server/apis/schedule/pkg/csv"
 	"github.com/jak103/powerplay/internal/server/apis/schedule/pkg/parser"
+	"github.com/jak103/powerplay/internal/server/services/auth"
 	"github.com/jak103/powerplay/internal/utils/responder"
 )
+
+func init() {
+	apis.RegisterHandler(fiber.MethodPost, "/schedule/ref", auth.Authenticated, handleRef)
+}
 
 type RefScheduleRow struct {
 	Start           string `csv:"Start Date and Time"`
