--- backend/internal/server/apis/schedule/ref_test.go
@@ -1,1 +1,0 @@
-package schedule
