+++ backend/internal/server/apis/schedule/rsvp/rsvp_test.go
@@ -1,0 +1,1 @@
+package rsvp
