--- backend/internal/server/apis/schedule/rsvp_test.go
@@ -1,1 +1,0 @@
-package schedule
